split reqeust
