+++ chain-signatures/node/src/protocol/request/mod.rs
@@ -34,6 +34,16 @@ use std::time::{Duration, Instant};
 use tokio::sync::{mpsc, watch, OwnedSemaphorePermit, Semaphore};
 use tokio::task::JoinHandle;
 
+mod limiter;
+mod metrics;
+mod organize;
+mod posit;
+mod state;
+mod task;
+
+use limiter::SignLimiter;
+use task::SignTask;
+
 /// The round interval to search for a proposer in the organizing phase.
 const ROUND_INTERVAL: usize = 512;
 
@@ -68,1116 +78,6 @@ pub enum Sign {
     Completion(SignId),
     Checkpoint(IndexedSignRequest),
     AbortChain(Chain),
-}
-
-#[derive(Debug)]
-pub enum SignLimitError {
-    Timeout,
-    Closed,
-}
-
-#[derive(Debug)]
-struct SignLimitState {
-    limit: usize,
-    debt: usize,
-}
-
-#[derive(Clone, Debug)]
-pub struct SignLimiter {
-    semaphore: Arc<Semaphore>,
-    state: Arc<RwLock<SignLimitState>>,
-}
-
-#[derive(Debug)]
-pub struct SignPermit {
-    permit: Option<OwnedSemaphorePermit>,
-    state: Arc<RwLock<SignLimitState>>,
-}
-
-impl SignLimiter {
-    pub fn new(limit: usize) -> Self {
-        Self {
-            semaphore: Arc::new(Semaphore::new(limit)),
-            state: Arc::new(RwLock::new(SignLimitState { limit, debt: 0 })),
-        }
-    }
-
-    /// Updates the limits for concurrent slots
-    #[cfg_attr(not(test), allow(dead_code))]
-    pub fn update(&self, new_limit: usize) {
-        let mut state = match self.state.write() {
-            Ok(state) => state,
-            Err(err) => {
-                tracing::error!(new_limit, ?err, "unable to update SignLimiter limits");
-                return;
-            }
-        };
-        let old_limit = std::mem::replace(&mut state.limit, new_limit);
-
-        // add more permits if the limit increased
-        if new_limit > old_limit {
-            let mut permits_to_add = new_limit - old_limit;
-            if permits_to_add > 0 {
-                let forgiven = permits_to_add.min(state.debt);
-                state.debt -= forgiven;
-                permits_to_add -= forgiven;
-                self.semaphore.add_permits(permits_to_add);
-            }
-            return;
-        }
-
-        // remove permits or add to a debt where when the permit is dropped, we forget it.
-        let permits_to_remove = old_limit - new_limit;
-        if permits_to_remove == 0 {
-            return;
-        }
-
-        let forgotten = self.semaphore.forget_permits(permits_to_remove);
-        if forgotten < permits_to_remove {
-            state.debt += permits_to_remove - forgotten;
-        }
-    }
-
-    /// Try to acquire a spot with a timeout just in case we do not receive the slot in time.
-    /// Returns a permit if successful, error otherwise.
-    pub async fn acquire(&self, timeout: Duration) -> Result<SignPermit, SignLimitError> {
-        let permit =
-            match tokio::time::timeout(timeout, self.semaphore.clone().acquire_owned()).await {
-                Ok(Ok(permit)) => permit,
-                // note, acquire error is effectively the same as closed.
-                Ok(Err(_acquire_err)) => return Err(SignLimitError::Closed),
-                Err(_timeout) => return Err(SignLimitError::Timeout),
-            };
-
-        Ok(SignPermit {
-            permit: Some(permit),
-            state: Arc::clone(&self.state),
-        })
-    }
-
-    pub fn limits(&self) -> usize {
-        match self.state.read() {
-            Ok(state) => state.limit,
-            Err(err) => {
-                tracing::error!(?err, "failed to acquire lock in SignLimiter::limits");
-                0
-            }
-        }
-    }
-}
-
-impl Drop for SignPermit {
-    fn drop(&mut self) {
-        let Some(permit) = self.permit.take() else {
-            return;
-        };
-        let mut state = match self.state.write() {
-            Ok(state) => state,
-            Err(err) => {
-                tracing::error!(?err, "failed to acquire lock in SignPermit drop");
-                return;
-            }
-        };
-        if state.debt > 0 {
-            state.debt -= 1;
-            permit.forget();
-        }
-    }
-}
-
-struct SignState {
-    round: usize,
-    indexed: IndexedSignRequest,
-    mesh_state: watch::Receiver<MeshState>,
-    /// Budget for the current organizing+posit attempt.
-    budget: TimeoutBudget,
-    permit: Option<SignPermit>,
-    /// The highest round sent by a peer
-    highest_seen_round: usize,
-    /// Posit message for `highest_seen_round` round.
-    ///
-    /// These are later processed, if the task reaches the `highest_seen_round`
-    /// as a deliberator. Proposers do not reprocess old messages. A valid peer
-    /// would not have sent a posit message before the proposer proposes.
-    ///
-    /// INVARIANT: All messages stored here are for `highest_seen_round`. Must
-    /// be cleared when `highest_seen_round` changes.
-    buffered_messages: VecDeque<SignTaskMessage>,
-    /// When Some, another group is already generating this signature.
-    /// The timestamp is when proposing can be resumed.
-    pause_proposing: Option<std::time::Instant>,
-}
-
-impl SignState {
-    fn new(indexed: IndexedSignRequest, mesh_state: watch::Receiver<MeshState>) -> Self {
-        Self {
-            round: 0,
-            indexed,
-            mesh_state,
-            budget: TimeoutBudget::new(ORGANIZE_POSIT_TIMEOUT),
-            permit: None,
-            highest_seen_round: 0,
-            buffered_messages: VecDeque::new(),
-            pause_proposing: None,
-        }
-    }
-
-    fn indexed(&self) -> &IndexedSignRequest {
-        &self.indexed
-    }
-
-    fn bump_round(&mut self) {
-        let prev_round = self.round;
-        self.round = std::cmp::max(self.round + 1, self.highest_seen_round);
-        // Reset the budget for the new attempt
-        self.budget.reset(ORGANIZE_POSIT_TIMEOUT);
-        self.permit = None;
-        tracing::debug!(prev_round, new_round = self.round, "bumped round");
-    }
-
-    /// When receiving posit message for future rounds, store them away until
-    /// that round is reached.
-    fn store_future_posit_message(&mut self, msg: SignTaskMessage) {
-        let SignTaskMessage::PositMessage {
-            round: peer_round, ..
-        } = msg;
-
-        if peer_round < self.highest_seen_round {
-            return;
-        }
-        if peer_round > self.highest_seen_round {
-            self.highest_seen_round = peer_round;
-            self.buffered_messages.clear();
-        }
-        self.buffered_messages.push_back(msg);
-    }
-
-    /// Remove a buffered message for processing, if there is one for the
-    /// current round.
-    fn take_buffered_posit_message(&mut self) -> Option<SignTaskMessage> {
-        if self.highest_seen_round == self.round {
-            self.buffered_messages.pop_front()
-        } else {
-            None
-        }
-    }
-}
-
-struct SignPositor {
-    proposer: Participant,
-    active: BTreeSet<Participant>,
-    presignature_id: PresignatureId,
-    presignature: Option<PresignatureReservation>,
-}
-
-struct SignGenerating {
-    proposer: Participant,
-    presignature_id: PresignatureId,
-    presignature: Option<PresignatureReservation>,
-    accepted_participants: Vec<Participant>,
-}
-
-enum SignPhase {
-    Organizing(SignOrganizer),
-    Posit(SignPositor),
-    Generating(SignGenerating),
-    Complete(Result<(), SignError>),
-}
-
-impl SignPhase {
-    async fn advance(
-        &mut self,
-        ctx: &mut SignTask,
-        state: &mut SignState,
-        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
-    ) -> SignPhase {
-        match self {
-            SignPhase::Organizing(phase) => phase.advance(ctx, state).await,
-            SignPhase::Posit(phase) => phase.advance(ctx, state, task_rx).await,
-            SignPhase::Generating(phase) => phase.advance(ctx, state, task_rx).await,
-            SignPhase::Complete(result) => SignPhase::Complete(*result),
-        }
-    }
-}
-
-struct SignOrganizer;
-
-impl SignOrganizer {
-    fn proposer_per_round(
-        round: usize,
-        participants: &[Participant],
-        entropy: &[u8; 32],
-    ) -> Participant {
-        let index = entropy[0] as usize + round;
-        participants[index % participants.len()]
-    }
-
-    /// Waits for threshold active participants to be present.
-    async fn wait_active(
-        &self,
-        ctx: &mut SignTask,
-        state: &mut SignState,
-        threshold: usize,
-    ) -> Option<BTreeSet<Participant>> {
-        let sign_id = ctx.sign_id;
-        let mut once = true;
-
-        loop {
-            let active_count = {
-                let active: BTreeSet<_> =
-                    state.mesh_state.borrow().active().keys().copied().collect();
-                if active.len() >= threshold {
-                    return Some(active);
-                }
-                active.len()
-            };
-
-            if once {
-                tracing::info!(
-                    ?sign_id,
-                    active_count,
-                    ?threshold,
-                    "waiting for enough active participants"
-                );
-                once = false;
-            }
-
-            if state.mesh_state.changed().await.is_err() {
-                return None;
-            }
-        }
-    }
-
-    async fn advance(&mut self, ctx: &mut SignTask, state: &mut SignState) -> SignPhase {
-        let sign_id = ctx.sign_id;
-        let threshold = ctx.governance.threshold;
-        let me = ctx.governance.me;
-        let entropy = state.indexed.args.entropy;
-        let participants = ctx
-            .governance
-            .participants
-            .iter()
-            .copied()
-            .collect::<Vec<_>>();
-
-        ctx.is_proposer.store(false, Ordering::Relaxed);
-
-        tracing::info!(?sign_id, round = ?state.round, "entering organizing phase");
-        let (active, proposer, is_proposer) = {
-            let Some(active) = self.wait_active(ctx, state, threshold).await else {
-                tracing::warn!(?sign_id, round = ?state.round, "no active participants, reorganizing");
-                state.bump_round();
-                return SignPhase::Organizing(SignOrganizer);
-            };
-
-            let max_rounds = state.round + ROUND_INTERVAL;
-            let (selected_round, proposer) = (state.round..max_rounds)
-                .map(|r| (r, Self::proposer_per_round(r, &participants, &entropy)))
-                .find(|(_, potential_proposer)| active.contains(potential_proposer))
-                .unwrap_or_else(|| {
-                    (
-                        max_rounds,
-                        *active
-                            .iter()
-                            .choose(&mut StdRng::from_seed(entropy))
-                            .unwrap(),
-                    )
-                });
-
-            state.round = selected_round;
-
-            // If proposing is paused (generating already ongoing), we act as if we weren't a proposer.
-            let skip_proposing = state
-                .pause_proposing
-                .is_some_and(|until| until > std::time::Instant::now());
-            let is_proposer = proposer == me && !skip_proposing;
-            ctx.is_proposer.store(is_proposer, Ordering::Relaxed);
-
-            tracing::info!(
-                ?sign_id,
-                round = selected_round,
-                ?proposer,
-                ?me,
-                is_proposer,
-                skip_proposing,
-                active_count = active.len(),
-                "organized: selected proposer"
-            );
-
-            (active, proposer, is_proposer)
-        };
-
-        if is_proposer {
-            let remaining = state.budget.remaining();
-            tracing::info!(
-                ?sign_id,
-                round = ?state.round,
-                timeout = ?remaining,
-                limit = ctx.limiter.limits(),
-                "proposer waiting for concurrency slot"
-            );
-
-            let permit = match ctx.limiter.acquire(remaining).await {
-                Ok(permit) => permit,
-                Err(SignLimitError::Timeout) => {
-                    tracing::warn!(
-                        ?sign_id,
-                        round = ?state.round,
-                        "proposer timeout waiting for concurrency slot, reorganizing"
-                    );
-                    state.bump_round();
-                    return SignPhase::Organizing(SignOrganizer);
-                }
-                Err(SignLimitError::Closed) => {
-                    tracing::error!(?sign_id, "proposer semaphore closed");
-                    return SignPhase::Complete(Err(SignError::Aborted));
-                }
-            };
-
-            state.permit = Some(permit);
-        } else {
-            state.permit = None;
-        }
-
-        let (presignature_id, presignature, active) = if is_proposer {
-            tracing::info!(?sign_id, round = ?state.round, "proposer waiting for presignature");
-            let active = active.iter().copied().collect::<Vec<_>>();
-            let remaining = state.budget.remaining();
-            let fetch = tokio::time::timeout(remaining, async {
-                // IDs that were found unsuitable this round (kept in redis, skipped next peek)
-                let mut local_skip: Vec<PresignatureId> = Vec::new();
-                loop {
-                    if let Some(reservation) = ctx.presignatures.peek_mine(&local_skip).await {
-                        let holders = reservation.holders();
-                        let participants = intersect_vec(&[holders, &active]);
-                        if participants.len() < ctx.governance.threshold {
-                            tracing::warn!(
-                                ?sign_id,
-                                id = reservation.id,
-                                ?holders,
-                                ?active,
-                                "skipping presignature due to inactive participants, returning to pool"
-                            );
-                            local_skip.push(reservation.id);
-                            // drop: in-memory reservation released, presignature stays in redis
-                            continue;
-                        }
-
-                        break (reservation, participants);
-                    }
-                    tokio::time::sleep(Duration::from_millis(500)).await;
-                }
-            })
-            .await;
-
-            let (reservation, participants) = match fetch {
-                Ok(value) => value,
-                Err(_) => {
-                    tracing::warn!(
-                        ?sign_id,
-                        round = ?state.round,
-                        "proposer timeout waiting for presignature, reorganizing"
-                    );
-                    state.bump_round();
-                    return SignPhase::Organizing(SignOrganizer);
-                }
-            };
-
-            let presignature_id = reservation.id;
-
-            tracing::info!(?sign_id, ?presignature_id, "proposer got presignature");
-
-            // broadcast to participants and let them reject if they don't have the presignature.
-            for &p in &participants {
-                if p == ctx.governance.me {
-                    continue;
-                }
-                ctx.msg
-                    .send(
-                        ctx.governance.me,
-                        p,
-                        PositMessage {
-                            id: PositProtocolId::Signature(sign_id, presignature_id, state.round),
-                            from: ctx.governance.me,
-                            action: PositAction::Propose,
-                        },
-                    )
-                    .await;
-            }
-
-            // Update active to only include participants that are in both the presignature and active set
-            let active = participants.into_iter().collect::<BTreeSet<_>>();
-            (presignature_id, Some(reservation), active)
-        } else {
-            (PresignatureId::default(), None, active)
-        };
-
-        SignPhase::Posit(SignPositor {
-            proposer,
-            active,
-            presignature_id,
-            presignature,
-        })
-    }
-}
-
-impl SignPositor {
-    /// Deliberator waits for the proposer to send a Propose message with a presignature_id.
-    async fn wait_propose(
-        ctx: &mut SignTask,
-        state: &mut SignState,
-        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
-        proposer: Participant,
-    ) -> Result<PresignatureId, SignPhase> {
-        let sign_id = ctx.sign_id;
-        let round = state.round;
-        let remaining = state.budget.remaining();
-        let outcome = tokio::time::timeout(remaining, async {
-            loop {
-                // Prioritize buffered messages, if any for the current round
-                let task_msg = match state.take_buffered_posit_message() {
-                    Some(buffered) => buffered,
-                    None => {
-                        let Some(task_msg) = task_rx.recv().await else {
-                            continue;
-                        };
-                        task_msg
-                    }
-                };
-
-                let SignTaskMessage::PositMessage {
-                    presignature_id,
-                    from,
-                    action,
-                    round: peer_round,
-                } = &task_msg;
-
-                // reject any messages with a different round than ours
-                //
-                // note: Rejecting messages of older rounds is always the right
-                // choice. But for newer messages, we could buffer them and try
-                // that round later. What we must not do is immediately jump to
-                // that higher round, or else any peer could force themselves to
-                // be the proposer every time.
-                if state.round > *peer_round {
-                    ctx.msg
-                        .send(
-                            ctx.governance.me,
-                            *from,
-                            PositMessage {
-                                id: PositProtocolId::Signature(
-                                    sign_id,
-                                    *presignature_id,
-                                    *peer_round,
-                                ),
-                                from: ctx.governance.me,
-                                action: PositAction::RejectWithReason(
-                                    PositRejectReason::InvalidRequest,
-                                ),
-                            },
-                        )
-                        .await;
-                    continue;
-                }
-
-                // Message can't be processed now but is crucial to make progress later.
-                // Note that we must first try and finish the current round and
-                // not immediately jump to that higher round. Otherwise, any peer
-                // could force themselves to be the proposer every time.
-                if state.round < *peer_round {
-                    tracing::info!(
-                        peer_round,
-                        my_round = state.round,
-                        "Storing message for future round, as deliberator",
-                    );
-                    state.store_future_posit_message(task_msg);
-                    continue;
-                }
-
-                if !matches!(action, PositAction::Propose) {
-                    tracing::warn!(
-                        round = peer_round,
-                        ?action,
-                        "Got unexpected posit message while waiting for propose"
-                    );
-                    continue;
-                }
-
-                if from == &proposer {
-                    tracing::info!(
-                        ?sign_id,
-                        ?presignature_id,
-                        ?from,
-                        "deliberator received Propose"
-                    );
-
-                    // Check if we have access to this presignature (in storage or generating)
-                    if !ctx.presignatures.contains(*presignature_id).await {
-                        tracing::warn!(
-                            ?sign_id,
-                            presignature_id,
-                            "deliberator does not have access to proposed presignature, rejecting"
-                        );
-                        ctx.msg
-                            .send(
-                                ctx.governance.me,
-                                proposer,
-                                PositMessage {
-                                    id: PositProtocolId::Signature(
-                                        sign_id,
-                                        *presignature_id,
-                                        state.round,
-                                    ),
-                                    from: ctx.governance.me,
-                                    action: PositAction::RejectWithReason(
-                                        PositRejectReason::MissingArtifact,
-                                    ),
-                                },
-                            )
-                            .await;
-                        continue;
-                    }
-
-                    break Ok(*presignature_id);
-                } else {
-                    tracing::warn!(
-                        ?sign_id,
-                        ?from,
-                        ?proposer,
-                        "received Propose from non-proposer, rejecting"
-                    );
-
-                    ctx.msg
-                        .send(
-                            ctx.governance.me,
-                            *from,
-                            PositMessage {
-                                id: PositProtocolId::Signature(
-                                    sign_id,
-                                    *presignature_id,
-                                    state.round,
-                                ),
-                                from: ctx.governance.me,
-                                action: PositAction::RejectWithReason(
-                                    PositRejectReason::InvalidRequest,
-                                ),
-                            },
-                        )
-                        .await;
-                }
-            }
-        })
-        .await;
-
-        let presignature_id = match outcome {
-            Ok(Ok(id)) => id,
-            Ok(Err(phase)) => return Err(phase),
-            Err(_) => {
-                tracing::warn!(
-                    ?sign_id,
-                    ?round,
-                    ?proposer,
-                    me=?ctx.governance.me,
-                    "deliberator timeout waiting for Propose, reorganizing"
-                );
-                state.bump_round();
-                return Err(SignPhase::Organizing(SignOrganizer));
-            }
-        };
-
-        // received propose, send Accept
-        ctx.msg
-            .send(
-                ctx.governance.me,
-                proposer,
-                PositMessage {
-                    id: PositProtocolId::Signature(sign_id, presignature_id, state.round),
-                    from: ctx.governance.me,
-                    action: PositAction::Accept,
-                },
-            )
-            .await;
-
-        Ok(presignature_id)
-    }
-
-    async fn advance(
-        &mut self,
-        ctx: &mut SignTask,
-        state: &mut SignState,
-        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
-    ) -> SignPhase {
-        let proposer = self.proposer;
-        let active = self.active.clone();
-        let mut presignature_id = self.presignature_id;
-        let presignature = self.presignature.take();
-
-        let sign_id = ctx.sign_id;
-        let round = state.round;
-        let is_proposer = proposer == ctx.governance.me;
-        let is_deliberator = !is_proposer;
-
-        tracing::info!(
-            ?sign_id,
-            ?presignature_id,
-            ?round,
-            is_proposer,
-            "entering posit phase"
-        );
-
-        if is_deliberator {
-            tracing::info!(
-                ?sign_id,
-                ?round,
-                ?proposer,
-                "deliberator waiting for Propose"
-            );
-
-            presignature_id = match Self::wait_propose(ctx, state, task_rx, proposer).await {
-                Ok(id) => id,
-                Err(phase) => return phase,
-            }
-        }
-
-        // GUARANTEE: at least threshold participants from organizing phase.
-        let posit_participants = active.iter().copied().collect::<Vec<_>>();
-        let mut counter = SinglePositCounter::new(ctx.governance.me, &posit_participants);
-
-        let mut remaining = state.budget.remaining();
-        if is_deliberator {
-            // We just sent an Accept to the proposer. The proposer might wait up to ACCEPT_POSIT_TIMEOUT
-            // to gather more accepts before sending Start.
-            // We must wait at least that long so we don't abandon the round after promising to participate,
-            // which would cause the proposer's generation phase to hang.
-            let min_wait = ACCEPT_POSIT_TIMEOUT + Duration::from_millis(500);
-            if remaining < min_wait {
-                remaining = min_wait;
-            }
-        }
-        let posit_deadline = tokio::time::sleep(remaining);
-        tokio::pin!(posit_deadline);
-        let accept_deadline = tokio::time::sleep(ACCEPT_POSIT_TIMEOUT);
-        tokio::pin!(accept_deadline);
-        let mut accept_deadline_reached = false;
-
-        let accepted_participants = loop {
-            tokio::select! {
-                Some(task_msg) = task_rx.recv() => {
-                    let SignTaskMessage::PositMessage { round: peer_round , ..} = task_msg;
-
-                    // Ignore messages for older rounds
-                    if state.round > peer_round {
-                        continue;
-                    }
-
-                    // Message can't be processed now but is crucial to make progress later.
-                    // Note that we must first try and finish the current round and
-                    // not immediately jump to that higher round. Otherwise, any peer
-                    // could force themselves to be the proposer every time.
-                    if state.round < peer_round {
-                        tracing::info!(
-                            peer_round,
-                            my_round = state.round,
-                            "Storing message for future round",
-                        );
-                        state.store_future_posit_message(task_msg);
-                        continue;
-                    }
-
-                    let SignTaskMessage::PositMessage { presignature_id: _, round: _peer_round, from, action } = task_msg;
-
-                    if is_deliberator {
-                        if let PositAction::Start(participants) = action {
-                            if from != proposer {
-                                tracing::warn!(?sign_id, ?round, ?from, ?proposer, "received Start from non-proposer, ignoring");
-                                continue;
-                            }
-
-                            if participants.len() < ctx.governance.threshold {
-                                tracing::warn!(
-                                    ?sign_id,
-                                    ?round,
-                                    "not enough start participants"
-                                );
-                                state.bump_round();
-                                return SignPhase::Organizing(SignOrganizer);
-                            }
-
-                            tracing::info!(?sign_id, participant = ?ctx.governance.me, ?participants, "deliberator received Start");
-                            break participants;
-                        }
-                    } else {
-                        if !counter.process_action(from, &action) {
-                            continue;
-                        }
-
-                        if counter.enough_rejects(ctx.governance.threshold) {
-                            let num_ongoing = counter.num_peers_with_ongoing_generation();
-                            if ctx.governance.participants.len().saturating_sub(num_ongoing) < ctx.governance.threshold {
-                                state.pause_proposing = Some(Instant::now() + Duration::from_millis(ctx.cfg.signature.generation_timeout));
-                                tracing::info!(
-                                    ?sign_id,
-                                    ?round,
-                                    resume=?state.pause_proposing,
-                                    "pausing proposer: peers already generating this signature"
-                                );
-                            } else {
-                                tracing::warn!(
-                                    ?sign_id,
-                                    ?round,
-                                    ?from,
-                                    "received enough REJECTs, reorganizing"
-                                );
-                            }
-                            if let Some(_reservation) = presignature {
-                                tracing::warn!(?sign_id, "returning presignature to pool due to REJECTs");
-                            }
-                            state.bump_round();
-                            return SignPhase::Organizing(SignOrganizer);
-                        }
-
-                        // Starting as soon as we have enough accepts leaves
-                        // participants accepting a bit later in a bad state.
-                        // They will try to become propose in later rounds,
-                        // wasting Presignatures, memory and CPU time.
-                        //
-                        // Instead, wait for at least the `accept_deadline`,
-                        // only nodes answer slower will be left out. This isn't
-                        // perfect but much better than always forcing nodes
-                        // into the bad state.
-                        let ready_to_go = counter.meets_totality() ||  accept_deadline_reached;
-                        if ready_to_go && counter.enough_accepts(ctx.governance.threshold) {
-                            let participants = Self::start_with_current_accepts(
-                                ctx,
-                                state,
-                                counter,
-                                sign_id,
-                                presignature_id
-                            ).await;
-                            break participants;
-                        }
-                    }
-                }
-                _ = &mut posit_deadline => {
-                    if is_proposer {
-                        tracing::warn!(
-                            ?sign_id,
-                            accepts = counter.accepts.len(),
-                            threshold = ctx.governance.threshold,
-                            ?round,
-                            "proposer posit deadline reached, expiring round"
-                        );
-                        if let Some(_reservation) = presignature {
-                            tracing::warn!(?sign_id, "returning presignature to pool due to proposer timeout");
-                        }
-                    } else {
-                        tracing::warn!(?sign_id, me=?ctx.governance.me, ?proposer, "deliberator posit timeout waiting for Start, reorganizing");
-                    }
-
-                    state.bump_round();
-                    return SignPhase::Organizing(SignOrganizer);
-                }
-                _ = &mut accept_deadline, if is_proposer && !accept_deadline_reached => {
-                    accept_deadline_reached = true;
-                    if counter.enough_accepts(ctx.governance.threshold) {
-                        let participants = Self::start_with_current_accepts(
-                            ctx,
-                            state,
-                            counter,
-                            sign_id,
-                            presignature_id
-                        ).await;
-                        break participants;
-                    }
-                }
-
-            }
-        };
-
-        SignPhase::Generating(SignGenerating {
-            proposer,
-            presignature_id,
-            presignature,
-            accepted_participants,
-        })
-    }
-
-    async fn start_with_current_accepts(
-        ctx: &SignTask,
-        state: &mut SignState,
-        counter: SinglePositCounter,
-        sign_id: SignId,
-        presignature_id: PresignatureId,
-    ) -> Vec<Participant> {
-        let participants = counter.accepts.into_iter().collect::<Vec<_>>();
-        tracing::info!(?sign_id, round=?state.round, me = ?ctx.governance.me, ?participants, "proposer broadcasting Start");
-
-        for &p in &participants {
-            if p == ctx.governance.me {
-                continue;
-            }
-            ctx.msg
-                .send(
-                    ctx.governance.me,
-                    p,
-                    PositMessage {
-                        id: PositProtocolId::Signature(sign_id, presignature_id, state.round),
-                        from: ctx.governance.me,
-                        action: PositAction::Start(participants.clone()),
-                    },
-                )
-                .await;
-        }
-        participants
-    }
-}
-
-impl SignGenerating {
-    async fn advance(
-        &mut self,
-        ctx: &SignTask,
-        state: &mut SignState,
-        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
-    ) -> SignPhase {
-        // We successfully committed to generating; future rounds should be unrestricted.
-        state.pause_proposing = None;
-
-        let sign_id = ctx.sign_id;
-        let round = state.round;
-
-        tracing::info!(
-            ?sign_id,
-            presignature_id = ?self.presignature_id,
-            participants = ?self.accepted_participants,
-            "posit complete, starting generation"
-        );
-
-        let presignature_pending = if let Some(reservation) = self.presignature.take() {
-            // Commit: actually remove from Redis now that posit succeeded and generation starts
-            match reservation.commit().await {
-                Some(taken) => PendingPresignature::Available(Box::new(taken)),
-                None => {
-                    tracing::warn!(
-                        ?sign_id,
-                        ?round,
-                        "failed to commit presignature reservation, reorganizing"
-                    );
-                    state.bump_round();
-                    return SignPhase::Organizing(SignOrganizer);
-                }
-            }
-        } else {
-            PendingPresignature::InStorage(
-                self.presignature_id,
-                self.proposer,
-                ctx.presignatures.clone(),
-            )
-        };
-
-        let gen_ctx = ctx.generate_ctx();
-        let generator = match SignGenerator::new(
-            &gen_ctx,
-            self.proposer,
-            state.indexed().clone(),
-            presignature_pending,
-            self.accepted_participants.clone(),
-        )
-        .await
-        {
-            Ok(gen) => gen,
-            Err(err) => {
-                tracing::warn!(
-                    ?sign_id,
-                    ?round,
-                    ?err,
-                    "failed to create generator, reorganizing"
-                );
-                state.bump_round();
-                return SignPhase::Organizing(SignOrganizer);
-            }
-        };
-
-        // Track that we've created a generator
-        crate::metrics::protocols::NUM_TOTAL_HISTORICAL_SIGNATURE_GENERATORS.inc();
-
-        match generator.run(&gen_ctx, task_rx).await {
-            Ok(()) => SignPhase::Complete(Ok(())),
-            Err(err) => {
-                tracing::warn!(
-                    ?sign_id,
-                    ?round,
-                    ?err,
-                    me=?ctx.governance.me,
-                    "signature generation failed, reorganizing"
-                );
-                state.bump_round();
-                SignPhase::Organizing(SignOrganizer)
-            }
-        }
-    }
-}
-
-/// Per-request accumulator for the three looping phases in `SignTask::run`:
-/// Organizing, Posit, Generating. Times are summed across attempts so each
-/// histogram observation covers the full request even when the state machine
-/// loops back. Indexing/AwaitingGeneration/Responding/Total are emitted
-/// elsewhere and ignored by `add`.
-///
-/// Additivity caveat: without governance pauses, all five stages sum to
-/// Total. Resharing or other transitions out of `Running` mid-request show
-/// idle time only in Total, so the equality holds as `<=` in that case.
-#[derive(Default, Debug, Clone, Copy, PartialEq, Eq)]
-struct PhaseDurations {
-    organizing: Duration,
-    posit: Duration,
-    generating: Duration,
-}
-
-impl PhaseDurations {
-    fn add(&mut self, step: SignRequestStep, elapsed: Duration) {
-        match step {
-            SignRequestStep::Organizing => self.organizing += elapsed,
-            SignRequestStep::Posit => self.posit += elapsed,
-            SignRequestStep::Generating => self.generating += elapsed,
-            // Emitted elsewhere; listed explicitly so adding a new variant
-            // forces a deliberate decision here.
-            SignRequestStep::Indexing
-            | SignRequestStep::AwaitingGeneration
-            | SignRequestStep::Responding
-            | SignRequestStep::Total => {}
-        }
-    }
-
-    fn emit(self, chain: Chain) {
-        record_request_latency(chain, SignRequestStep::Organizing, "ok", self.organizing);
-        record_request_latency(chain, SignRequestStep::Posit, "ok", self.posit);
-        record_request_latency(chain, SignRequestStep::Generating, "ok", self.generating);
-    }
-}
-
-struct SignTask {
-    governance: GovernanceInfo,
-    sign_id: SignId,
-    presignatures: PresignatureStorage,
-    msg: MessageChannel,
-    rpc: RpcChannel,
-
-    // TODO: will be used in the future when we move requests channels
-    // into the backlog.
-    #[allow(dead_code)]
-    backlog: Backlog,
-
-    cfg: ProtocolConfig,
-    contract: ContractStateWatcher,
-    is_proposer: Arc<AtomicBool>,
-    limiter: SignLimiter,
-    node_account_id: near_account_id::AccountId,
-}
-
-impl SignTask {
-    async fn run(
-        mut self,
-        indexed: IndexedSignRequest,
-        mesh_state: watch::Receiver<MeshState>,
-        mut task_rx: mpsc::Receiver<SignTaskMessage>,
-    ) -> Result<(), SignError> {
-        let sign_id = self.sign_id;
-        tracing::info!(?sign_id, governance = ?self.governance, "signature task starting...");
-
-        let mut state = SignState::new(indexed, mesh_state);
-        let mut phase = SignPhase::Organizing(SignOrganizer);
-        let mut contract_watcher = self.contract.clone();
-
-        // NOTE: even if we start the SignTask while in resharing, we will not advance
-        // since we won't in the running state
-        let mut is_running = self.governance.is_running;
-
-        // Sum per-phase time across loop attempts; emit on Complete(Ok) only.
-        let mut durations = PhaseDurations::default();
-
-        loop {
-            let phase_start = Instant::now();
-            let current_phase_step = match &phase {
-                SignPhase::Organizing(_) => Some(SignRequestStep::Organizing),
-                SignPhase::Posit(_) => Some(SignRequestStep::Posit),
-                SignPhase::Generating(_) => Some(SignRequestStep::Generating),
-                SignPhase::Complete(_) => None,
-            };
-
-            tokio::select! {
-                Some(contract_state) = contract_watcher.next_state() => {
-                    // `phase.advance` was cancelled. Attribute its partial
-                    // elapsed time only if `is_running` was true; otherwise
-                    // the branch was gated off and the iteration was idle.
-                    if is_running {
-                        if let Some(step) = current_phase_step {
-                            durations.add(step, phase_start.elapsed());
-                        }
-                    }
-                    is_running = self.refresh_governance(&contract_state);
-                    if is_running {
-                        // we're back into running, reset to SignOrganizer
-                        phase = SignPhase::Organizing(SignOrganizer);
-                    } else {
-                        tracing::info!(
-                            ?sign_id,
-                            gov = ?self.governance,
-                            "signature task paused waiting for running governance"
-                        );
-                    }
-                }
-                // This branch in tokio::select will get cancelled since the future for next contract
-                // state is reached first. This effectively pauses this branch from executing and
-                // further advancing the signature organization/positing/generation flow.
-                new_phase = phase.advance(&mut self, &mut state, &mut task_rx), if is_running => {
-                    if let Some(step) = current_phase_step {
-                        durations.add(step, phase_start.elapsed());
-                        if matches!(&new_phase, SignPhase::Organizing(_)) {
-                            SIGN_REQUEST_LOOPS
-                                .with_label_values(&[
-                                    state.indexed().chain.as_str(),
-                                    step.as_str(),
-                                ])
-                                .inc();
-                        }
-                    }
-
-                    match new_phase {
-                        SignPhase::Complete(result) => {
-                            if result.is_ok() {
-                                durations.emit(state.indexed().chain);
-                            }
-                            return result;
-                        }
-                        new_phase => phase = new_phase,
-                    }
-                }
-            };
-        }
-    }
-
-    /// Refresh the governance info from the contract state, returning whether we are
-    /// in a running state with valid governance info after the refresh.
-    fn refresh_governance(&mut self, contract_state: &ProtocolState) -> bool {
-        if let Some(governance) = contract_state.governance(&self.node_account_id) {
-            self.governance = governance;
-        } else {
-            self.governance.is_running = false;
-        }
-        self.governance.is_running
-    }
-
-    /// Snapshot the fields the generation protocol needs.
-    fn generate_ctx(&self) -> GenerateCtx {
-        GenerateCtx {
-            governance: self.governance.clone(),
-            msg: self.msg.clone(),
-            rpc: self.rpc.clone(),
-            backlog: self.backlog.clone(),
-            cfg: self.cfg.clone(),
-            node_account_id: self.node_account_id.clone(),
-        }
-    }
 }
 
 pub struct SignatureSpawner {
@@ -1555,224 +455,10 @@ pub fn organize_posit_timeout() -> Duration {
 mod tests {
     use super::*;
     use crate::protocol::contract::primitives::{ParticipantInfo, Participants};
-    use crate::protocol::contract::{ResharingContractState, RunningContractState};
     use crate::protocol::presignature::Presignature;
-    use crate::protocol::ProtocolState;
 
     use cait_sith::protocol::Participant;
     use deadpool_redis::Runtime;
-
-    #[test]
-    fn sign_task_refreshes_and_pauses_on_resharing() {
-        let account_id: near_account_id::AccountId = "p-0".parse().unwrap();
-        let mut participants = Participants::default();
-        participants.insert(&Participant::from(0), ParticipantInfo::new(0));
-
-        let governance = GovernanceInfo {
-            me: Participant::from(0),
-            threshold: 1,
-            epoch: 0,
-            public_key: k256::AffinePoint::default(),
-            participants: [Participant::from(0)].into_iter().collect(),
-            is_running: true,
-        };
-
-        let redis_cfg = deadpool_redis::Config::from_url("redis://127.0.0.1/");
-        let pool = redis_cfg.create_pool(Some(Runtime::Tokio1)).unwrap();
-        let presignatures = Presignature::storage(&pool, &account_id);
-        let (_inbox, _outbox, msg_channel) = MessageChannel::new();
-        let (rpc_tx, _rpc_rx) = mpsc::channel(1);
-        let rpc_channel = RpcChannel { tx: rpc_tx };
-        let (contract, _tx) = ContractStateWatcher::with_running(
-            &account_id,
-            k256::AffinePoint::default(),
-            1,
-            participants.clone(),
-        );
-
-        let mut sign_task = SignTask {
-            governance: governance.clone(),
-            sign_id: SignId::new([0u8; 32]),
-            presignatures,
-            msg: msg_channel,
-            rpc: rpc_channel,
-            backlog: Backlog::new(),
-            cfg: ProtocolConfig::default(),
-            contract,
-            is_proposer: Arc::new(AtomicBool::new(false)),
-            limiter: SignLimiter::new(1),
-            node_account_id: account_id.clone(),
-        };
-
-        let initial = RunningContractState {
-            epoch: 0,
-            public_key: k256::AffinePoint::default(),
-            participants: participants.clone(),
-            candidates: Default::default(),
-            join_votes: Default::default(),
-            leave_votes: Default::default(),
-            threshold: 1,
-        };
-        assert!(sign_task.refresh_governance(&ProtocolState::Running(initial)));
-        assert_eq!(sign_task.governance.epoch, 0);
-        assert_eq!(sign_task.governance.threshold, 1);
-        assert_eq!(sign_task.governance.me, Participant::from(0));
-
-        let resharing = ResharingContractState {
-            old_epoch: 0,
-            old_participants: participants.clone(),
-            new_participants: participants.clone(),
-            threshold: 1,
-            public_key: k256::AffinePoint::default(),
-            finished_votes: Default::default(),
-            cancel_votes: Default::default(),
-        };
-
-        // refreshing here should yield false where we are no longer in running.
-        assert!(!sign_task.refresh_governance(&ProtocolState::Resharing(resharing)));
-
-        let running = RunningContractState {
-            epoch: 1,
-            public_key: k256::AffinePoint::default(),
-            participants,
-            candidates: Default::default(),
-            join_votes: Default::default(),
-            leave_votes: Default::default(),
-            threshold: 1,
-        };
-
-        assert!(sign_task.refresh_governance(&ProtocolState::Running(running)));
-        assert_eq!(sign_task.governance.epoch, 1);
-        assert_eq!(sign_task.governance.threshold, 1);
-        assert_eq!(sign_task.governance.me, Participant::from(0));
-    }
-
-    #[test]
-    fn phase_durations_sum_per_phase_across_attempts() {
-        // Simulates a request that loops Organizing -> Posit -> Organizing ->
-        // Posit -> Generating before completing. Each `add` mirrors one
-        // iteration of the SignTask::run phase loop.
-        let mut d = PhaseDurations::default();
-        d.add(SignRequestStep::Organizing, Duration::from_millis(100));
-        d.add(SignRequestStep::Posit, Duration::from_millis(200));
-        // back-edge: Posit failed and we re-entered Organizing
-        d.add(SignRequestStep::Organizing, Duration::from_millis(50));
-        d.add(SignRequestStep::Posit, Duration::from_millis(150));
-        d.add(SignRequestStep::Generating, Duration::from_millis(500));
-
-        assert_eq!(d.organizing, Duration::from_millis(150));
-        assert_eq!(d.posit, Duration::from_millis(350));
-        assert_eq!(d.generating, Duration::from_millis(500));
-    }
-
-    #[test]
-    fn phase_durations_ignore_steps_not_part_of_sign_task() {
-        // These variants are not part of SignTask: Indexing is emitted by
-        // the indexer modules, AwaitingGeneration by SignatureSpawner::
-        // handle_request, and Responding/Total by rpc.rs. Passing any to
-        // `add` must be a no-op or they would be double-counted on
-        // multichain_sign_request_latency_sec.
-        let mut d = PhaseDurations::default();
-        d.add(SignRequestStep::Indexing, Duration::from_millis(100));
-        d.add(
-            SignRequestStep::AwaitingGeneration,
-            Duration::from_millis(200),
-        );
-        d.add(SignRequestStep::Responding, Duration::from_millis(300));
-        d.add(SignRequestStep::Total, Duration::from_millis(400));
-
-        assert_eq!(d, PhaseDurations::default());
-    }
-
-    #[test]
-    fn phase_durations_preserve_additivity_invariant() {
-        // This represents the case where we have 2 attempts:
-        // 1st one fails at posit, then starts from organizing again and finishes
-        let inputs = [
-            (SignRequestStep::Organizing, Duration::from_millis(40)),
-            (SignRequestStep::Posit, Duration::from_millis(120)),
-            (SignRequestStep::Organizing, Duration::from_millis(35)),
-            (SignRequestStep::Posit, Duration::from_millis(95)),
-            (SignRequestStep::Generating, Duration::from_millis(710)),
-        ];
-        let expected: Duration = inputs.iter().map(|(_, d)| *d).sum();
-
-        let mut d = PhaseDurations::default();
-        for (step, elapsed) in inputs {
-            d.add(step, elapsed);
-        }
-
-        assert_eq!(d.organizing + d.posit + d.generating, expected);
-    }
-
-    #[tokio::test]
-    async fn sign_limiter_times_out_at_limit() {
-        let semaphore = SignLimiter::new(1);
-        let permit = semaphore
-            .acquire(Duration::from_millis(10))
-            .await
-            .expect("first acquire should succeed");
-
-        let second = semaphore.acquire(Duration::from_millis(10)).await;
-        assert!(matches!(second, Err(SignLimitError::Timeout)));
-
-        drop(permit);
-
-        let third = semaphore.acquire(Duration::from_millis(10)).await;
-        assert!(third.is_ok());
-    }
-
-    #[tokio::test]
-    async fn sign_limiter_applies_reduced_limit_on_release() {
-        let semaphore = SignLimiter::new(2);
-        let first = semaphore
-            .acquire(Duration::from_millis(10))
-            .await
-            .expect("first acquire should succeed");
-        let second = semaphore
-            .acquire(Duration::from_millis(10))
-            .await
-            .expect("second acquire should succeed");
-
-        semaphore.update(1);
-
-        drop(first);
-
-        let third = semaphore.acquire(Duration::from_millis(10)).await;
-        assert!(matches!(third, Err(SignLimitError::Timeout)));
-
-        drop(second);
-
-        let fourth = semaphore.acquire(Duration::from_millis(10)).await;
-        assert!(fourth.is_ok());
-    }
-
-    /// Some integrations tests rely on this. If this changes, you probably also
-    /// have to update the tests.
-    #[test]
-    fn proposer_order_for_seed_0() {
-        // SHA256([0,0,0,0])[0] = 223.
-        // With 3 participants:
-        //   round 0: (223 + 0) % 3 = 1 -> participant 1
-        //   round 1: (223 + 1) % 3 = 2 -> participant 2
-        //   round 2: (223 + 2) % 3 = 0 -> participant 0
-        let participants: Vec<Participant> = (0u32..3).map(Participant::from).collect();
-        // entropy[0] = SHA256([0,0,0,0])[0] = 0xdf = 223; only the first byte matters.
-        let mut entropy = [0u8; 32];
-        entropy[0] = 0xdf;
-        assert_eq!(
-            SignOrganizer::proposer_per_round(0, &participants, &entropy),
-            Participant::from(1)
-        );
-        assert_eq!(
-            SignOrganizer::proposer_per_round(1, &participants, &entropy),
-            Participant::from(2)
-        );
-        assert_eq!(
-            SignOrganizer::proposer_per_round(2, &participants, &entropy),
-            Participant::from(0)
-        );
-    }
 
     #[tokio::test]
     async fn test_abort_chain_dead_ids_lifecycle() {
