+++ chain-signatures/node/src/protocol/request/posit.rs
@@ -1,0 +1,422 @@
+use super::organize::SignOrganizer;
+use super::state::SignState;
+use super::task::{SignGenerating, SignPhase};
+use super::*;
+
+pub struct SignPositor {
+    pub proposer: Participant,
+    pub active: BTreeSet<Participant>,
+    pub presignature_id: PresignatureId,
+    pub presignature: Option<PresignatureReservation>,
+}
+
+impl SignPositor {
+    /// Deliberator waits for the proposer to send a Propose message with a presignature_id.
+    async fn wait_propose(
+        ctx: &mut SignTask,
+        state: &mut SignState,
+        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
+        proposer: Participant,
+    ) -> Result<PresignatureId, SignPhase> {
+        let sign_id = ctx.sign_id;
+        let round = state.round;
+        let remaining = state.budget.remaining();
+        let outcome = tokio::time::timeout(remaining, async {
+            loop {
+                // Prioritize buffered messages, if any for the current round
+                let task_msg = match state.take_buffered_posit_message() {
+                    Some(buffered) => buffered,
+                    None => {
+                        let Some(task_msg) = task_rx.recv().await else {
+                            continue;
+                        };
+                        task_msg
+                    }
+                };
+
+                let SignTaskMessage::PositMessage {
+                    presignature_id,
+                    from,
+                    action,
+                    round: peer_round,
+                } = &task_msg;
+
+                // reject any messages with a different round than ours
+                //
+                // note: Rejecting messages of older rounds is always the right
+                // choice. But for newer messages, we could buffer them and try
+                // that round later. What we must not do is immediately jump to
+                // that higher round, or else any peer could force themselves to
+                // be the proposer every time.
+                if state.round > *peer_round {
+                    ctx.msg
+                        .send(
+                            ctx.governance.me,
+                            *from,
+                            PositMessage {
+                                id: PositProtocolId::Signature(
+                                    sign_id,
+                                    *presignature_id,
+                                    *peer_round,
+                                ),
+                                from: ctx.governance.me,
+                                action: PositAction::RejectWithReason(
+                                    PositRejectReason::InvalidRequest,
+                                ),
+                            },
+                        )
+                        .await;
+                    continue;
+                }
+
+                // Message can't be processed now but is crucial to make progress later.
+                // Note that we must first try and finish the current round and
+                // not immediately jump to that higher round. Otherwise, any peer
+                // could force themselves to be the proposer every time.
+                if state.round < *peer_round {
+                    tracing::info!(
+                        peer_round,
+                        my_round = state.round,
+                        "Storing message for future round, as deliberator",
+                    );
+                    state.store_future_posit_message(task_msg);
+                    continue;
+                }
+
+                if !matches!(action, PositAction::Propose) {
+                    tracing::warn!(
+                        round = peer_round,
+                        ?action,
+                        "Got unexpected posit message while waiting for propose"
+                    );
+                    continue;
+                }
+
+                if from == &proposer {
+                    tracing::info!(
+                        ?sign_id,
+                        ?presignature_id,
+                        ?from,
+                        "deliberator received Propose"
+                    );
+
+                    // Check if we have access to this presignature (in storage or generating)
+                    if !ctx.presignatures.contains(*presignature_id).await {
+                        tracing::warn!(
+                            ?sign_id,
+                            presignature_id,
+                            "deliberator does not have access to proposed presignature, rejecting"
+                        );
+                        ctx.msg
+                            .send(
+                                ctx.governance.me,
+                                proposer,
+                                PositMessage {
+                                    id: PositProtocolId::Signature(
+                                        sign_id,
+                                        *presignature_id,
+                                        state.round,
+                                    ),
+                                    from: ctx.governance.me,
+                                    action: PositAction::RejectWithReason(
+                                        PositRejectReason::MissingArtifact,
+                                    ),
+                                },
+                            )
+                            .await;
+                        continue;
+                    }
+
+                    break Ok(*presignature_id);
+                } else {
+                    tracing::warn!(
+                        ?sign_id,
+                        ?from,
+                        ?proposer,
+                        "received Propose from non-proposer, rejecting"
+                    );
+
+                    ctx.msg
+                        .send(
+                            ctx.governance.me,
+                            *from,
+                            PositMessage {
+                                id: PositProtocolId::Signature(
+                                    sign_id,
+                                    *presignature_id,
+                                    state.round,
+                                ),
+                                from: ctx.governance.me,
+                                action: PositAction::RejectWithReason(
+                                    PositRejectReason::InvalidRequest,
+                                ),
+                            },
+                        )
+                        .await;
+                }
+            }
+        })
+        .await;
+
+        let presignature_id = match outcome {
+            Ok(Ok(id)) => id,
+            Ok(Err(phase)) => return Err(phase),
+            Err(_) => {
+                tracing::warn!(
+                    ?sign_id,
+                    ?round,
+                    ?proposer,
+                    me=?ctx.governance.me,
+                    "deliberator timeout waiting for Propose, reorganizing"
+                );
+                state.bump_round();
+                return Err(SignPhase::Organizing(SignOrganizer));
+            }
+        };
+
+        // received propose, send Accept
+        ctx.msg
+            .send(
+                ctx.governance.me,
+                proposer,
+                PositMessage {
+                    id: PositProtocolId::Signature(sign_id, presignature_id, state.round),
+                    from: ctx.governance.me,
+                    action: PositAction::Accept,
+                },
+            )
+            .await;
+
+        Ok(presignature_id)
+    }
+
+    pub async fn advance(
+        &mut self,
+        ctx: &mut SignTask,
+        state: &mut SignState,
+        task_rx: &mut mpsc::Receiver<SignTaskMessage>,
+    ) -> SignPhase {
+        let proposer = self.proposer;
+        let active = self.active.clone();
+        let mut presignature_id = self.presignature_id;
+        let presignature = self.presignature.take();
+
+        let sign_id = ctx.sign_id;
+        let round = state.round;
+        let is_proposer = proposer == ctx.governance.me;
+        let is_deliberator = !is_proposer;
+
+        tracing::info!(
+            ?sign_id,
+            ?presignature_id,
+            ?round,
+            is_proposer,
+            "entering posit phase"
+        );
+
+        if is_deliberator {
+            tracing::info!(
+                ?sign_id,
+                ?round,
+                ?proposer,
+                "deliberator waiting for Propose"
+            );
+
+            presignature_id = match Self::wait_propose(ctx, state, task_rx, proposer).await {
+                Ok(id) => id,
+                Err(phase) => return phase,
+            }
+        }
+
+        // GUARANTEE: at least threshold participants from organizing phase.
+        let posit_participants = active.iter().copied().collect::<Vec<_>>();
+        let mut counter = SinglePositCounter::new(ctx.governance.me, &posit_participants);
+
+        let mut remaining = state.budget.remaining();
+        if is_deliberator {
+            // We just sent an Accept to the proposer. The proposer might wait up to ACCEPT_POSIT_TIMEOUT
+            // to gather more accepts before sending Start.
+            // We must wait at least that long so we don't abandon the round after promising to participate,
+            // which would cause the proposer's generation phase to hang.
+            let min_wait = ACCEPT_POSIT_TIMEOUT + Duration::from_millis(500);
+            if remaining < min_wait {
+                remaining = min_wait;
+            }
+        }
+        let posit_deadline = tokio::time::sleep(remaining);
+        tokio::pin!(posit_deadline);
+        let accept_deadline = tokio::time::sleep(ACCEPT_POSIT_TIMEOUT);
+        tokio::pin!(accept_deadline);
+        let mut accept_deadline_reached = false;
+
+        let accepted_participants = loop {
+            tokio::select! {
+                Some(task_msg) = task_rx.recv() => {
+                    let SignTaskMessage::PositMessage { round: peer_round , ..} = task_msg;
+
+                    // Ignore messages for older rounds
+                    if state.round > peer_round {
+                        continue;
+                    }
+
+                    // Message can't be processed now but is crucial to make progress later.
+                    // Note that we must first try and finish the current round and
+                    // not immediately jump to that higher round. Otherwise, any peer
+                    // could force themselves to be the proposer every time.
+                    if state.round < peer_round {
+                        tracing::info!(
+                            peer_round,
+                            my_round = state.round,
+                            "Storing message for future round",
+                        );
+                        state.store_future_posit_message(task_msg);
+                        continue;
+                    }
+
+                    let SignTaskMessage::PositMessage { presignature_id: _, round: _peer_round, from, action } = task_msg;
+
+                    if is_deliberator {
+                        if let PositAction::Start(participants) = action {
+                            if from != proposer {
+                                tracing::warn!(?sign_id, ?round, ?from, ?proposer, "received Start from non-proposer, ignoring");
+                                continue;
+                            }
+
+                            if participants.len() < ctx.governance.threshold {
+                                tracing::warn!(
+                                    ?sign_id,
+                                    ?round,
+                                    "not enough start participants"
+                                );
+                                state.bump_round();
+                                return SignPhase::Organizing(SignOrganizer);
+                            }
+
+                            tracing::info!(?sign_id, participant = ?ctx.governance.me, ?participants, "deliberator received Start");
+                            break participants;
+                        }
+                    } else {
+                        if !counter.process_action(from, &action) {
+                            continue;
+                        }
+
+                        if counter.enough_rejects(ctx.governance.threshold) {
+                            let num_ongoing = counter.num_peers_with_ongoing_generation();
+                            if ctx.governance.participants.len().saturating_sub(num_ongoing) < ctx.governance.threshold {
+                                state.pause_proposing = Some(Instant::now() + Duration::from_millis(ctx.cfg.signature.generation_timeout));
+                                tracing::info!(
+                                    ?sign_id,
+                                    ?round,
+                                    resume=?state.pause_proposing,
+                                    "pausing proposer: peers already generating this signature"
+                                );
+                            } else {
+                                tracing::warn!(
+                                    ?sign_id,
+                                    ?round,
+                                    ?from,
+                                    "received enough REJECTs, reorganizing"
+                                );
+                            }
+                            if let Some(_reservation) = presignature {
+                                tracing::warn!(?sign_id, "returning presignature to pool due to REJECTs");
+                            }
+                            state.bump_round();
+                            return SignPhase::Organizing(SignOrganizer);
+                        }
+
+                        // Starting as soon as we have enough accepts leaves
+                        // participants accepting a bit later in a bad state.
+                        // They will try to become propose in later rounds,
+                        // wasting Presignatures, memory and CPU time.
+                        //
+                        // Instead, wait for at least the `accept_deadline`,
+                        // only nodes answer slower will be left out. This isn't
+                        // perfect but much better than always forcing nodes
+                        // into the bad state.
+                        let ready_to_go = counter.meets_totality() ||  accept_deadline_reached;
+                        if ready_to_go && counter.enough_accepts(ctx.governance.threshold) {
+                            let participants = Self::start_with_current_accepts(
+                                ctx,
+                                state,
+                                counter,
+                                sign_id,
+                                presignature_id
+                            ).await;
+                            break participants;
+                        }
+                    }
+                }
+                _ = &mut posit_deadline => {
+                    if is_proposer {
+                        tracing::warn!(
+                            ?sign_id,
+                            accepts = counter.accepts.len(),
+                            threshold = ctx.governance.threshold,
+                            ?round,
+                            "proposer posit deadline reached, expiring round"
+                        );
+                        if let Some(_reservation) = presignature {
+                            tracing::warn!(?sign_id, "returning presignature to pool due to proposer timeout");
+                        }
+                    } else {
+                        tracing::warn!(?sign_id, me=?ctx.governance.me, ?proposer, "deliberator posit timeout waiting for Start, reorganizing");
+                    }
+
+                    state.bump_round();
+                    return SignPhase::Organizing(SignOrganizer);
+                }
+                _ = &mut accept_deadline, if is_proposer && !accept_deadline_reached => {
+                    accept_deadline_reached = true;
+                    if counter.enough_accepts(ctx.governance.threshold) {
+                        let participants = Self::start_with_current_accepts(
+                            ctx,
+                            state,
+                            counter,
+                            sign_id,
+                            presignature_id
+                        ).await;
+                        break participants;
+                    }
+                }
+
+            }
+        };
+
+        SignPhase::Generating(SignGenerating {
+            proposer,
+            presignature_id,
+            presignature,
+            accepted_participants,
+        })
+    }
+
+    async fn start_with_current_accepts(
+        ctx: &SignTask,
+        state: &mut SignState,
+        counter: SinglePositCounter,
+        sign_id: SignId,
+        presignature_id: PresignatureId,
+    ) -> Vec<Participant> {
+        let participants = counter.accepts.into_iter().collect::<Vec<_>>();
+        tracing::info!(?sign_id, round=?state.round, me = ?ctx.governance.me, ?participants, "proposer broadcasting Start");
+
+        for &p in &participants {
+            if p == ctx.governance.me {
+                continue;
+            }
+            ctx.msg
+                .send(
+                    ctx.governance.me,
+                    p,
+                    PositMessage {
+                        id: PositProtocolId::Signature(sign_id, presignature_id, state.round),
+                        from: ctx.governance.me,
+                        action: PositAction::Start(participants.clone()),
+                    },
+                )
+                .await;
+        }
+        participants
+    }
+}
